+++ pkg/rtc/transport.go
@@ -283,11 +283,6 @@ type PCTransport struct {
 	signalStateCheckTimer     *time.Timer
 	currentOfferIceCredential string // ice user:pwd, for publish side ice restart checking
 	pendingRestartIceOffer    *webrtc.SessionDescription
-
-	// CLOSE-DEBUG-CLEANUP
-	iceGatheringState   atomic.Value // webrtc.ICEGatheringState
-	iceConnectionState  atomic.Value // webrtc.ICEConnectionState
-	peerConnectionState atomic.Value // webrtc.PeerConnectionState
 }
 
 type TransportParams struct {
@@ -519,9 +514,6 @@ func NewPCTransport(params TransportParams) (*PCTransport, error) {
 		connectionDetails:        types.NewICEConnectionDetails(params.Transport, params.Logger),
 		lastNegotiate:            time.Now(),
 	}
-	t.iceGatheringState.Store(webrtc.ICEGatheringStateUnknown)
-	t.iceConnectionState.Store(webrtc.ICEConnectionStateUnknown)
-	t.peerConnectionState.Store(webrtc.PeerConnectionStateUnknown)
 	t.localOfferId.Store(uint32(rand.Intn(1<<8) + 1))
 
 	bwe, err := t.createPeerConnection()
@@ -770,7 +762,6 @@ func (t *PCTransport) setConnectedAt(at time.Time) bool {
 }
 
 func (t *PCTransport) onICEGatheringStateChange(state webrtc.ICEGatheringState) {
-	t.iceGatheringState.Store(state)
 	t.params.Logger.Debugw("ice gathering state change", "state", state.String())
 	if state != webrtc.ICEGatheringStateComplete {
 		return
@@ -802,7 +793,6 @@ func (t *PCTransport) handleConnectionFailed(forceShortConn bool) {
 }
 
 func (t *PCTransport) onICEConnectionStateChange(state webrtc.ICEConnectionState) {
-	t.iceConnectionState.Store(state)
 	t.params.Logger.Debugw("ice connection state change", "state", state.String())
 	switch state {
 	case webrtc.ICEConnectionStateConnected:
@@ -814,7 +804,6 @@ func (t *PCTransport) onICEConnectionStateChange(state webrtc.ICEConnectionState
 }
 
 func (t *PCTransport) onPeerConnectionStateChange(state webrtc.PeerConnectionState) {
-	t.peerConnectionState.Store(state)
 	t.params.Logger.Debugw("peer connection state change", "state", state.String())
 	switch state {
 	case webrtc.PeerConnectionStateConnected:
@@ -1501,54 +1490,16 @@ func (t *PCTransport) Close() {
 		return
 	}
 
-	var eventsQueueDone atomic.Bool
-	var streamAllocatorStopped atomic.Bool
-	var pacerStopped atomic.Bool
-	var reliableDataChannelClosed atomic.Bool
-	var lossyDataChannelClosed atomic.Bool
-	var unlabeledDataChannelClosed atomic.Bool
-	var peerConnectionClosed atomic.Bool
-	time.AfterFunc(time.Minute, func() { // CLOSE-DEBUG-CLEANUP
-		if !eventsQueueDone.Load() || !streamAllocatorStopped.Load() || !pacerStopped.Load() || !reliableDataChannelClosed.Load() || !lossyDataChannelClosed.Load() || !unlabeledDataChannelClosed.Load() || !peerConnectionClosed.Load() {
-			t.lock.Lock()
-			iceStartedAt := t.iceStartedAt
-			iceConnectedAt := t.iceConnectedAt
-			iceStats := t.mayFailedICEStats
-			t.lock.Unlock()
-
-			t.params.Logger.Infow(
-				"transport close timeout",
-				"eventsQueueDone", eventsQueueDone.Load(),
-				"streamAllocatorStopped", streamAllocatorStopped.Load(),
-				"pacerStopped", pacerStopped.Load(),
-				"reliableDataChannelClosed", reliableDataChannelClosed.Load(),
-				"lossyDataChannelClosed", lossyDataChannelClosed.Load(),
-				"unlabeledDataChannelClosed", unlabeledDataChannelClosed.Load(),
-				"peerConnectionClosed", peerConnectionClosed.Load(),
-				"iceStartedAt", iceStartedAt,
-				"iceConnectedAt", iceConnectedAt,
-				"iceGatheringState", t.iceGatheringState.Load().(webrtc.ICEGatheringState).String(),
-				"iceConnectionState", t.iceConnectionState.Load().(webrtc.ICEConnectionState).String(),
-				"peerConnectionState", t.peerConnectionState.Load().(webrtc.PeerConnectionState).String(),
-				"iceStats", iceCandidatePairStatsEncoder{iceStats},
-				"iceConnectionInfo", t.GetICEConnectionInfo(),
-				"connectionType", t.connectionDetails.GetConnectionType(),
-			)
-		}
-	})
-
 	<-t.eventsQueue.Stop()
-	eventsQueueDone.Store(true)
 	t.clearSignalStateCheckTimer()
 
 	if t.streamAllocator != nil {
 		t.streamAllocator.Stop()
 	}
-	streamAllocatorStopped.Store(true)
+
 	if t.pacer != nil {
 		t.pacer.Stop()
 	}
-	pacerStopped.Store(true)
 
 	t.clearConnTimer()
 
@@ -1562,13 +1513,11 @@ func (t *PCTransport) Close() {
 		t.reliableDC.Close()
 		t.reliableDC = nil
 	}
-	reliableDataChannelClosed.Store(true)
 
 	if t.lossyDC != nil {
 		t.lossyDC.Close()
 		t.lossyDC = nil
 	}
-	lossyDataChannelClosed.Store(true)
 
 	if t.dataTrackDC != nil {
 		t.dataTrackDC.Close()
@@ -1579,11 +1528,11 @@ func (t *PCTransport) Close() {
 		dc.Close()
 	}
 	t.unlabeledDataChannels = nil
-	unlabeledDataChannelClosed.Store(true)
 	t.lock.Unlock()
 
-	_ = t.pc.Close()
-	peerConnectionClosed.Store(true)
+	if err := t.pc.Close(); err != nil {
+		t.params.Logger.Warnw("unclean close of peer connection", err)
+	}
 
 	t.outputAndClearICEStats()
 }