+++ pkg/service/roommanager.go
@@ -338,6 +338,7 @@ func (r *RoomManager) StartSession(
 						Leave: leave,
 					},
 				})
+				prometheus.IncrementParticipantRtcCanceled(1)
 				return errors.New("could not restart closed participant")
 			}
 
@@ -395,6 +396,7 @@ func (r *RoomManager) StartSession(
 				Leave: leave,
 			},
 		})
+		prometheus.IncrementParticipantRtcCanceled(1)
 		return errors.New("could not restart participant")
 	}